Vision LLM/requirements.txt
streamlit
torch
transformers
pillow
pandas
scikit-learn
